Add repository functions and run them in service

+++ src/main/java/Repository/DepartmentRepository.java
@@ -3,5 +3,5 @@ package Repository;
 import Model.*;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DepartmentRepository extends CrudRepository<DepartmentsEntity, Long> {
+public interface DepartmentRepository extends CrudRepository<DepartmentsEntity, Integer> {
 }
